\name{DJIA}
\alias{DJIA}
\docType{data}
\title{
Dow Jones industrial average
}
\description{
The Dow Jones industrial average (DJIA) data set.
}
\usage{data("DJIA")}
\format{

  \describe{
    \item{\code{GAS.F}}{a numeric vector}
    \item{\code{Nikkei.F}}{a numeric vector}
    \item{\code{NZD}}{a numeric vector}
    \item{\code{silver.F}}{a numeric vector}
    \item{\code{RUSSELL.F}}{a numeric vector}
    \item{\code{S.P.F}}{a numeric vector}
    \item{\code{CHF}}{a numeric vector}
    \item{\code{Dollar.index.F}}{a numeric vector}
    \item{\code{Dollar.index}}{a numeric vector}
    \item{\code{wheat.F}}{a numeric vector}
    \item{\code{XAG}}{a numeric vector}
    \item{\code{XAU}}{a numeric vector}
  }
}
\details{
The data set comes from the Dow Jones industrial average (PSA)  data of 96 patients collected by Stanford University Medical Center. These patients all underwent radical prostatectomy.
}
\source{
The Stanford University Medical Center.
}
\references{
NA
}
\examples{
data(DJIA)
## maybe str(DJIA) ; plot(DJIA) ...
}
\keyword{datasets}
