\name{ISE}
\alias{ISE}
\docType{data}
\title{
Istanbul Stock Exchange
}
\description{
The Istanbul Stock Exchange (ISE) data set.
}
\usage{data("ISE")}
\format{

  \describe{
    \item{\code{ISE}}{a numeric vector}
    \item{\code{SP}}{a numeric vector}
    \item{\code{DAX}}{a numeric vector}
    \item{\code{FTSE}}{a numeric vector}
    \item{\code{NIKKEI}}{a numeric vector}
    \item{\code{BOVESPA}}{a numeric vector}
    \item{\code{EU}}{a numeric vector}
    \item{\code{EM}}{a numeric vector}
  }
}
\details{
The data set comes from the Istanbul Stock Exchange (ISE)  data of 96 patients collected by Stanford University Medical Center. These patients all underwent radical prostatectomy.
}
\source{
The Stanford University Medical Center.
}
\references{
NA
}
\examples{
data(ISE)
## maybe str(ISE) ; plot(ISE) ...
}
\keyword{datasets}

