\name{SECI}
\alias{SECI}
\docType{data}
\title{
New York Stock Exchange Composite Index
}
\description{
The New York Stock Exchange Composite Index SECI(SECI) data set.
}
\usage{data("SECI")}
\format{

  \describe{
    \item{\code{ GBP}}{a numeric vector}
    \item{\code{JPY}}{a numeric vector}
    \item{\code{CAD}}{a numeric vector}
    \item{\code{AAPL}}{a numeric vector}
    \item{\code{AMZN }}{a numeric vector}
    \item{\code{ GE}}{a numeric vector}
    \item{\code{JPM}}{a numeric vector}
    \item{\code{ MSFT}}{a numeric vector}
    \item{\code{WFC }}{a numeric vector}
    \item{\code{ XOM  }}{a numeric vector}
    \item{\code{FCHI }}{a numeric vector}
    \item{\code{FTSE }}{a numeric vector}
    \item{\code{GDAXI }}{a numeric vector}
  }
}
\details{
The data set comes from the prostate specific antigen (PSA)  data of 96 patients collected by Stanford University Medical Center. These patients all underwent radical prostatectomy.
}
\source{
The Stanford University Medical Center.
}
\references{
NA
}
\examples{
data(SECI)
## maybe str(SECI) ; plot(SECI) ...
}
\keyword{datasets}
