\name{SPP}
\alias{SPP}
\docType{data}
\title{
Stock Portfolio Performance
}
\description{
The Stock Portfolio Performance (SPP) data set.
}
\usage{data("SPP")}
\format{

  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
    \item{\code{X5}}{a numeric vector}
    \item{\code{X6}}{a numeric vector}
    \item{\code{X7}}{a numeric vector}
    \item{\code{X8}}{a numeric vector}
    \item{\code{X9}}{a numeric vector}
    \item{\code{X10}}{a numeric vector}
  }
}
\details{
The data set comes from the Stock Portfolio Performance (SPP)  data of 96 patients collected by Stanford University Medical Center. These patients all underwent radical prostatectomy.
}
\source{
The Stanford University Medical Center.
}
\references{
NA
}
\examples{
data(SPP)
## maybe str(SPP) ; plot(SPP) ...
}
\keyword{datasets}
